Validate that the script is running with a supported python version
I think the minimum veresion is actually 3.9, and that's only due to the
`type | type` syntax introduced in the last commit. Without that, it would
likely run on python 3.8. Still, 3.8 will be EOL in less than 6 months (2024-10)
so we might as well require a more recent version. With this version requirement in place, we support Ubuntu 22.04, Fedora 38, Debian 12
(bookworm), and other distros shipping python 3.10 or later. This is a
reasonable support window since python 3.10 was released on 2021-10-04.